Scoreboard no longer displays or shows empty scoreboards

+++ MPAi/Cores/Scoreboard/MPAiSoundScoreboardLoader.cs
@@ -26,6 +26,10 @@ namespace MPAi.Cores.Scoreboard
 
         public static void SaveScoreboard(MPAiSoundScoreBoard scoreboard)
         {
+            if (scoreboard.IsEmpty())
+            {
+                return;
+            }
             if (File.Exists(SpeakScoreboardFileAddress(scoreboard.User)))
             {
                 File.Delete(SpeakScoreboardFileAddress(scoreboard.User));
@@ -37,6 +41,10 @@ namespace MPAi.Cores.Scoreboard
                     sw.WriteLine("<Scoreboard>");
                     foreach (MPAiSoundScoreBoardSession session in scoreboard.Sessions)
                     {
+                        if (session.IsEmpty())
+                        {
+                            continue;
+                        }
                         sw.WriteLine("<Session>");
                         sw.WriteLine("<Date>");
                         sw.WriteLine(session.DateAndTime);
